Fix a subtle bug with discrete2explicit
Occasionally, an MDP may feature a state which is inaccessible from any other state (e.g. hidden behind a wall of terminal states). In such a case, there will be no transition to indicate the state-reward for this state. We arbitrarily leave the state-reward as 0 (this state will never be encountered, so it doesn't matter what value it takes on).