Support specifying a minimum path length in `log_partition` and partially in `sw_maxent_irl`
At this time, this is only implemented for `sw_maxent_irl(..., nll_only=True)` because I can't be bothered figuring out the indexing for the backward message passing variable and the marginal calculations. If you want to use a minimum path length for optimization purposes, consider using 2-point gradient estimation.